+++ pypacscrawler/executor.py
@@ -1,12 +1,11 @@
 import shlex
 import subprocess
 
-from typing import List, Dict
+from typing import List, Dict, Tuple
 from pypacscrawler.dicom import get_results
 
 
-def run(query):
-    # type: (str) -> (List[Dict[str, str]], int)
+def run(query: str) ->Tuple[List[Dict[str, str]], int]:
     """
     Executes a findscu query and parses the result
     :param query: findscu query
@@ -15,7 +14,7 @@ def run(query):
     """
     print(query)
     cmd = shlex.split(query)
-    completed = subprocess.run(cmd, stderr=subprocess.PIPE)
+    completed = subprocess.run(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
     lines = completed.stderr.decode('latin1').splitlines()
     result = get_results(lines)
     return result, len(result)