+++ pypacscrawler/time.py
@@ -1,4 +1,3 @@
-import datetime
 from datetime import timedelta, datetime